Gracefully handle negative-valued Encoding extensions
Back in 2012 libass added a custom Encoding=-1 extension to enable
automatic BiDi base direction instead of the usual LTR default.
The assumption/hope presumably was that since all valid font encodings
are positive VSFilter would just gracefully ignore them. However, this
is unfortuantely not the case.
When passed to GDI the charSet is cast to a byte and will wrap around to
a valid value, severely restricting which fonts can be used.

Searching through a large pile of subtitles the only hits for negative
values in the Encoding field were from generally corrupted files.
It thus appears so far fortunately nobody used the extension in released
fiels (only transient conversion from other  formats during playback)
and also nobody is relying on the wraparound in VSFilter.

To minimise problems and carve out the [-2, INT_MIN] space for
potential future extensions common to VSFilters and libass, gracefully
ignore them and use the default charset which is the most lenient in
font selection.
(This makes libass’ -1 slightly less broken in VSFilter and future
 common extension will continue to use the default charset for font
 selection)